Bug - SSH known hosts not compatible with Ansible
Ansible does SSH known host lookups via the FQDN.  While this appeared
to be recorded as such in the ssh_known_hosts file, closer examination
revealed that not to be true.  What looked like the FQDN was actually
that plus a suffix of an underscore and a key type.  It appears that
Puppet's sshkey resource type requires using the first host alias as
unique key ID for its own purposes.  That's acceptable, but our usage
resulted in the file not having the FQDN present for use by ssh.  This
revision adds the FQDN as a separate, additional alias so that Ansible
or any other use case where the FQDN is the ssh target will be able to
leverage these known hosts.